# backend/openapi.json
{
  "components": {
    "responses": {
...
      "UnauthorizedError": {
        "content": {
          "application/json": {
            "examples": {
              "authorization_header_missing": {
                "summary": "Authorization header missing",
                "value": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "Authorization header missing"
                  },
                  "success": false
                }
              },
              "invalid_token": {
                "summary": "Invalid or expired token",
                "value": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "Invalid or expired token"
                  },
                  "success": false
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Unauthorized - missing or invalid token"
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "details": {
                "type": "object"
              },
              "message": {
                "type": "string"
              }
            },
            "type": "object"
          },
          "success": {
            "example": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "SummaryResponse": {
        "properties": {
          "data": {
            "properties": {
              "expense_count": {
                "type": "integer"
              },
              "income_count": {
                "type": "integer"
              },
              "net_balance": {
                "description": "Decimal as string",
                "type": "string"
              },
              "total_expense": {
                "description": "Decimal as string",
                "type": "string"
              },
              "total_income": {
                "description": "Decimal as string",
                "type": "string"
              }
            },
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "Transaction": {
        "properties": {
          "amount": {
            "description": "Decimal as string",
            "type": "string"
          },
          "category": {
            "nullable": true,
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "date": {
            "format": "date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "merchant_name": {
            "nullable": true,
            "type": "string"
          },
          "transaction_type": {
            "enum": [
              "income",
              "expense"
            ],
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
      "TransactionCreate": {
        "properties": {
          "amount": {
            "description": "Decimal amount (e.g., '125.50')",
            "type": "string"
          },
          "category": {
            "maxLength": 100,
            "type": "string"
          },
          "date": {
            "format": "date",
            "type": "string"
          },
          "merchant_name": {
            "maxLength": 255,
            "type": "string"
          },
          "transaction_type": {
            "enum": [
              "income",
              "expense"
            ],
            "type": "string"
          }
        },
        "required": [
          "transaction_type",
          "amount",
          "date"
        ],
        "type": "object"
      },
      "TransactionListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Transaction"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          },
          "meta": {
            "properties": {
              "page": {
                "type": "integer"
              },
              "per_page": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "total_pages": {
                "type": "integer"
              }
            },
            "type": "object"
          },
          "success": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "TransactionResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Transaction"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "TransactionUpdate": {
        "properties": {
          "amount": {
            "description": "Decimal amount (e.g., '125.50')",
            "type": "string"
          },
          "category": {
            "maxLength": 100,
            "type": "string"
          },
          "date": {
            "format": "date",
            "type": "string"
          },
          "merchant_name": {
            "maxLength": 255,
            "type": "string"
          },
          "transaction_type": {
            "enum": [
              "income",
              "expense"
            ],
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Auth0 JWT access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "support@digitalfinance.com",
      "name": "Digital Finance Tracker Team"
    },
    "description": "\n## AI-Powered Personal Finance Management API\n\nThis API provides endpoints for managing personal finances including:\n- **Authentication** - Auth0-based user authentication\n- **Users** - User profile and settings management\n- **Transactions** - CRUD operations for financial transactions\n\n### Authentication\nAll protected endpoints require a valid Auth0 JWT token in the Authorization header:\n```\nAuthorization: Bearer <access_token>\n```\n\n### Response Format\nMost API endpoints follow a standard format:\n```json\n{\n    \"success\": true,\n    \"data\": {...},\n    \"message\": \"Success message\"\n}\n```\n\n### Special Cases\n**Exceptions**:\n- \"/\" and \"/health\" return unwrapped objects (no success/data/message).\n\nError responses:\n```json\n{\n    \"success\": false,\n    \"error\": {\n        \"code\": \"ERROR_CODE\",\n        \"message\": \"Error description\",\n        \"details\": {...}\n    }\n}\n```\n        ",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "Digital Finance Tracker API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/transactions": {
      "get": {
        "description": "Get paginated list of user's transactions",
        "parameters": [
          {
            "description": "Page number",
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "type": "integer"
            }
          },
          {
            "description": "Items per page (max 100)",
            "in": "query",
            "name": "per_page",
            "schema": {
              "default": 20,
              "maximum": 100,
              "type": "integer"
            }
          },
          {
            "description": "Filter by transaction type",
            "in": "query",
            "name": "transaction_type",
            "schema": {
              "enum": [
                "income",
                "expense"
              ],
              "type": "string"
            }
          },
          {
            "description": "Filter by category",
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter from date (YYYY-MM-DD)",
            "in": "query",
            "name": "start_date",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "description": "Filter to date (YYYY-MM-DD)",
            "in": "query",
            "name": "end_date",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "description": "Sort field",
            "in": "query",
            "name": "sort_by",
            "schema": {
              "default": "date",
              "enum": [
                "date",
                "amount",
                "created_at"
              ],
              "type": "string"
            }
          },
          {
            "description": "Sort order",
            "in": "query",
            "name": "sort_order",
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionListResponse"
                }
              }
            },
            "description": "List of transactions"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Transactions",
        "tags": [
          "Transactions"
        ]
      },
      "post": {
        "description": "Create a new transaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            },
            "description": "Transaction created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create Transaction",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/summary": {
      "get": {
        "description": "Get summary statistics for user's transactions",
        "parameters": [
          {
            "description": "Start date for summary (YYYY-MM-DD)",
            "in": "query",
            "name": "start_date",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "description": "End date for summary (YYYY-MM-DD)",
            "in": "query",
            "name": "end_date",
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResponse"
                }
              }
            },
            "description": "Transaction summary"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Transaction Summary",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{id}": {
      "delete": {
        "description": "Delete a transaction",
        "parameters": [
          {
            "description": "Transaction UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Transaction deleted"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Transaction not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete Transaction",
        "tags": [
          "Transactions"
        ]
      },
      "get": {
        "description": "Get a single transaction by ID",
        "parameters": [
          {
            "description": "Transaction UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            },
            "description": "Transaction details"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Transaction not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Transaction",
        "tags": [
          "Transactions"
        ]
      },
      "patch": {
        "description": "Update an existing transaction",
        "parameters": [
          {
            "description": "Transaction UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            },
            "description": "Transaction updated"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Transaction not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update Transaction",
        "tags": [
          "Transactions"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Development server",
      "url": "http://localhost:8000"
    }
  ],
...
}